Creating a new branch specifically for the Mayfly
This branch modifies the SDI-12 code to remove several of the ports such
that it will interact with other Mayfly libraries.
